Filtered nduckdb hook instead of period duckdb hook
