+++ packages/frontend/src/model/duckdb/types/use-duckdb-filtered-input.type.ts
@@ -1,0 +1,5 @@
+export type useDuckDbFilteredInput = {
+  periodId?: string;
+  startDate?: string;
+  endDate?: string;
+};
